# cmd/api/server/openapi.yaml
openapi: 3.1.0
info:
  title: Code Search API
  description: |
    A self-hosted code search and replace platform with support for batch operations across repositories.

    ## Features
    - Fast indexed search across all repositories (powered by Zoekt)
    - CLI for bulk operations (search & replace with MR creation)
    - GitHub & GitLab integration (clone, sync, create PRs/MRs)

    ## Authentication
    API requests can be authenticated using a Bearer token in the Authorization header.
  version: 1.0.0
  contact:
    name: Code Search Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://codesearch.example.com
    description: Production server
...
paths:
...
  /api/v1/connections/{id}:
    get:
      tags:
        - Connections
      summary: Get connection
      operationId: getConnection
      parameters:
        - $ref: "#/components/parameters/ConnectionId"
      responses:
        "200":
          description: Connection details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "404":
          description: Connection not found

    put:
      tags:
        - Connections
      summary: Update connection
      operationId: updateConnection
      parameters:
        - $ref: "#/components/parameters/ConnectionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConnectionRequest"
      responses:
        "200":
          description: Connection updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "404":
          description: Connection not found

    delete:
      tags:
        - Connections
      summary: Delete connection
      operationId: deleteConnection
      parameters:
        - $ref: "#/components/parameters/ConnectionId"
      responses:
        "204":
          description: Connection deleted
        "404":
          description: Connection not found
...
  /api/v1/connections/{id}/stats:
    get:
      tags:
        - Connections
      summary: Get connection statistics
      description: Returns detailed statistics for a specific connection including repo counts and indexing status.
      operationId: getConnectionStats
      parameters:
        - $ref: "#/components/parameters/ConnectionId"
      responses:
        "200":
          description: Connection statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionStats"
        "404":
          description: Connection not found

  # Replace Endpoints
...
components:
...
  parameters:
...
    ConnectionId:
      name: id
      in: path
      required: true
      description: Connection ID
      schema:
        type: integer
        format: int64
...
  schemas:
...
    Connection:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: GitHub Production
        type:
          type: string
          enum: [github, github_enterprise, gitlab, gitea, bitbucket]
        url:
          type: string
          format: uri
          example: https://github.com
        created_at:
          type: string
          format: date-time
...
    UpdateConnectionRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [github, github_enterprise, gitlab, gitea, bitbucket]
        url:
          type: string
          format: uri
        token:
          type: string
          format: password
...
    ConnectionStats:
      type: object
      properties:
        total_repos:
          type: integer
        indexed_repos:
          type: integer
        pending_repos:
          type: integer
        failed_repos:
          type: integer
        last_sync:
          type: string
          format: date-time

    # Queue Stats Schema
...